mcp_server.py
"""
MCP Server - Request to Standard
Expone herramientas para ser consumidas por agente orquestador

Ejecutar en modo stdio:
    python mcp_server.py

Ejecutar en modo HTTP:
    uvicorn mcp_server:mcp_app --port 8001

NOTA: Este es un placeholder para desarrollo futuro.
El MCP SDK se integrará completamente cuando se construya el agente orquestador.
"""
from typing import Optional, Literal
from custom_logging import get_logger

logger = get_logger(__name__)
# TODO: Descomentar cuando se instale MCP SDK
# from mcp.server import Server
# from mcp.types import Tool, TextContent

# Placeholder: servidor MCP básico
class MCPServer:
    """Servidor MCP para Request to Standard"""

    def __init__(self):
        self.name = "request-to-standard"
        self.version = "0.1.0"

    async def standardize_to_rag1(self, file_path: str) -> dict:
        """
        Tool: Estandarizar datos a formato RAG1

        Args:
            file_path: Ruta al archivo CSV/XLSX

        Returns:
            Datos estandarizados en formato RAG1
        """
        from src.core.pipeline import StandardizationPipeline

        # Leer archivo
        with open(file_path, 'rb') as f:
            file_content = f.read()

        filename = file_path.split('/')[-1]
        file_size = len(file_content)

        # Procesar con RAG1
        pipeline = StandardizationPipeline()
        result = await pipeline.process(
            file_content=file_content,
            filename=filename,
            file_size=file_size,
            target_rag="rag1"
        )

        return result.model_dump()

    async def standardize_to_rag2(self, file_path: str) -> dict:
        """
        Tool: Estandarizar datos a formato RAG2

        Args:
            file_path: Ruta al archivo CSV/XLSX

        Returns:
            Datos estandarizados en formato RAG2
        """
        from src.core.pipeline import StandardizationPipeline

        # Leer archivo
        with open(file_path, 'rb') as f:
            file_content = f.read()

        filename = file_path.split('/')[-1]
        file_size = len(file_content)

        # Procesar con RAG2
        pipeline = StandardizationPipeline()
        result = await pipeline.process(
            file_content=file_content,
            filename=filename,
            file_size=file_size,
            target_rag="rag2"
        )

        return result.model_dump()

    async def analyze_structure_tool(self, file_path: str) -> dict:
        """
        Tool: Analizar estructura de archivo

        Args:
            file_path: Ruta al archivo

        Returns:
            Análisis de estructura básica
        """
        from src.core.ingestion import DataIngestion
        from src.core.cleaning import DataCleaning
        from src.core.normalization import DataNormalization

        # Leer archivo
        with open(file_path, 'rb') as f:
            file_content = f.read()

        filename = file_path.split('/')[-1]

        # Procesar hasta normalización
        ingestion = DataIngestion()
        df = await ingestion.ingest(file_content, filename)

        cleaning = DataCleaning()
        df_clean = await cleaning.clean(df)

        normalization = DataNormalization()
        df_normalized = await normalization.normalize(df_clean)

        return {
            "filename": filename,
            "rows": len(df),
            "columns": list(df.columns),
            "dtypes": {col: str(dtype) for col, dtype in df_normalized.dtypes.items()},
            "sample": df_normalized.head(3).to_dict(orient='records')
        }

    async def get_schemas_tool(self) -> dict:
        """
        Tool: Obtener esquemas RAG disponibles

        Returns:
            Esquemas de RAG 1 y RAG 2
        """
        from src.models.rag1_schema import RAG1Schema
        from src.models.rag2_schema import RAG2Schema

        return {
            "rag1": RAG1Schema.model_json_schema(),
            "rag2": RAG2Schema.model_json_schema()
        }


# Ejemplo de integración MCP (cuando esté disponible el SDK)
"""
# Implementación completa con MCP SDK:

from mcp.server import Server
from mcp.types import Tool

server = Server("request-to-standard")

@server.tool()
async def standardize_to_rag1(file_path: str) -> str:
    '''Estandariza datos a formato RAG1 (documentos estructurados)'''
    mcp_server = MCPServer()
    result = await mcp_server.standardize_to_rag1(file_path)
    return json.dumps(result, indent=2)

@server.tool()
async def standardize_to_rag2(file_path: str) -> str:
    '''Estandariza datos a formato RAG2 (servicios/tickets)'''
    mcp_server = MCPServer()
    result = await mcp_server.standardize_to_rag2(file_path)
    return json.dumps(result, indent=2)

@server.tool()
async def analyze_structure(file_path: str) -> str:
    '''Analiza estructura de archivo'''
    mcp_server = MCPServer()
    result = await mcp_server.analyze_structure_tool(file_path)
    return json.dumps(result, indent=2)

@server.tool()
async def get_schemas() -> str:
    '''Obtiene esquemas RAG disponibles'''
    mcp_server = MCPServer()
    result = await mcp_server.get_schemas_tool()
    return json.dumps(result, indent=2)

# Ejecutar servidor
if __name__ == "__main__":
    import asyncio
    asyncio.run(server.run())
"""


# Por ahora: servidor simple para testing
def main():
    """Punto de entrada para MCP server"""
    logger.info(f"MCP Server: request-to-standard v0.1.0")
    logger.info("Modo: Placeholder (en desarrollo)")
    logger.info("\nTools disponibles:")
    logger.info("  - standardize_to_rag1: Estandarizar datos a RAG1")
    logger.info("  - standardize_to_rag2: Estandarizar datos a RAG2")
    logger.info("  - analyze_structure: Analizar estructura de archivo")
    logger.info("  - get_schemas: Obtener esquemas RAG")
    logger.info("\nNOTA: El orquestador decidirá qué tool usar (rag1 o rag2)")
    logger.info("Para integración completa, espere implementación del agente orquestador.")


if __name__ == "__main__":
    main()